Moved some types around
