--- src/components/Field/index.ts
@@ -1,1 +1,0 @@
-export * from './Field.types';
