+++ src/components/ValidationWrapper/ValidationWrapper.types.tsx
@@ -1,6 +1,5 @@
 import { TValidator, TAsyncValidator } from '../../validators';
-import { TValidateMethod, TResetMethod, TUpdateMethod, IValidationState } from '../../hooks';
-import { TBasicFieldValue } from '../Field';
+import { TValidateMethod, TResetMethod, TUpdateMethod, IValidationState, TBasicFieldValue } from '../../hooks';
 
 /**
  * Properties of a component that is wrapped
