+++ src/components/index.ts
@@ -4,8 +4,6 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-
-export * from './Field';
 export * from './FieldError';
 export * from './FieldGroup';
 export * from './FieldLine';
