+++ src/components/withField/index.ts
@@ -1,1 +1,2 @@
 export * from './withField';
+export * from './withField.types';
