+++ src/components/withValidation/withValidation.types.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { TBasicFieldValue } from '../Field';
+import { TBasicFieldValue } from '../../hooks';
 import { IValidationProp } from '../ValidationWrapper';
 
 /**
